restructure previous code, comment out intermediate scattering function
