+++ cpp/diffraction/StaticStructureFactorDebye.cc
@@ -35,6 +35,8 @@ float k_min_center_to_lower_edge(unsigned int bins, float k_min, float k_max)
 
 StaticStructureFactorDebye::StaticStructureFactorDebye(unsigned int bins, float k_max, float k_min)
     : StaticStructureFactor(bins, k_max_center_to_upper_edge(bins, k_min, k_max),
+                            k_min_center_to_lower_edge(bins, k_min, k_max)),
+    StructureFactor(bins, k_max_center_to_upper_edge(bins, k_min, k_max),
                             k_min_center_to_lower_edge(bins, k_min, k_max))
 {
     if (bins == 0)
