+++ freud/_diffraction.pxd
@@ -9,32 +9,37 @@ cimport freud.util
 from freud.util cimport vec3
 
 
-cdef extern from "StaticStructureFactor.h" namespace "freud::diffraction":
-    cdef cppclass StaticStructureFactor:
-        const freud.util.ManagedArray[float] &getStructureFactor()
+cdef extern from "StructureFactor.h" namespace "freud::diffraction":
+    cdef cppclass StructureFactor:
         const vector[float] getBinEdges() const
         const vector[float] getBinCenters() const
         float getMinValidK() const
 
 cdef extern from "StaticStructureFactorDebye.h" namespace "freud::diffraction":
-    cdef cppclass StaticStructureFactorDebye(StaticStructureFactor):
+    cdef cppclass StaticStructureFactorDebye(StructureFactor):
         StaticStructureFactorDebye(unsigned int, float, float) except +
+        const freud.util.ManagedArray[float] &getStructureFactor()
         void accumulate(const freud._locality.NeighborQuery*,
                         const vec3[float]*, unsigned int, unsigned int) except +
         void reset()
 
-cdef extern from "StaticStructureFactorDirect.h" namespace "freud::diffraction":
-    cdef cppclass StaticStructureFactorDirect(StaticStructureFactor):
-        StaticStructureFactorDirect(unsigned int, float, float, unsigned int) except +
-        void accumulate(const freud._locality.NeighborQuery*,
-                        const vec3[float]*, unsigned int, unsigned int) except +
-        void reset()
+cdef extern from "StructureFactorDirect.h" namespace "freud::diffraction":
+    cdef cppclass StructureFactorDirect(StructureFactor):
         unsigned int getNumSampledKPoints() const
         vector[vec3[float]] getKPoints() const
 
+cdef extern from "StaticStructureFactorDirect.h" namespace "freud::diffraction":
+    cdef cppclass StaticStructureFactorDirect(StructureFactorDirect):
+        StaticStructureFactorDirect(unsigned int, float, float, unsigned int) except +
+        const freud.util.ManagedArray[float] &getStructureFactor()
+        void accumulate(const freud._locality.NeighborQuery*,
+                        const vec3[float]*, unsigned int, unsigned int) except +
+        void reset()
+
+"""
 cdef extern from "IntermediateScattering.h" namespace "freud::diffraction":
     cdef cppclass IntermediateScattering(StaticStructureFactorDirect):
         IntermediateScattering(unsigned int, float, float, unsigned int) except +
         const freud.util.ManagedArray[float] &getSelfFunction() const
         const freud.util.ManagedArray[float] &getDistinctFunction() const
-
+"""
